Add explicit type conversions

+++ src/libscotch/vdgraph_separate_df.c
@@ -1,4 +1,4 @@
-/* Copyright 2007,2008,2023 IPB, Universite de Bordeaux, INRIA & CNRS
+/* Copyright 2007,2008,2023,2024 IPB, Universite de Bordeaux, INRIA & CNRS
 **
 ** This file is part of the Scotch software package for static mapping,
 ** graph partitioning and sparse matrix ordering.
@@ -43,7 +43,7 @@
 /**   DATES      : # Version 5.1  : from : 05 nov 2007     **/
 /**                                 to   : 09 nov 2008     **/
 /**                # Version 7.0  : from : 20 jan 2023     **/
-/**                                 to   : 20 jan 2023     **/
+/**                                 to   : 09 aug 2024     **/
 /**                                                        **/
 /************************************************************/
 
@@ -191,8 +191,8 @@ const VdgraphSeparateDfParam * const  paraptr)    /*+ Method parameters +*/
   }
 
   ovflval    = 0;
-  cdifval    = paraptr->cdifval;
-  cremval    = paraptr->cremval;
+  cdifval    = (float) paraptr->cdifval;
+  cremval    = (float) paraptr->cremval;
   edgegsttax = grafptr->s.edgegsttax;
   for (passnum = 0; ; ) {                         /* For all passes         */
     if (ovflval == 0) {                           /* If no overflow occured */
